refactor(Storybook): Name the anatomy model file for what it holds
It was called anatomy.ts only because pageAnatomy.ts collides with
PageAnatomy.tsx on a case-insensitive filesystem, which TypeScript rejects.
model.ts says what the file is: the shape of a page's anatomy and the reading of
it from a story, with no rendering.
